requirements.txt
flask
mysqlclient==2.1.1
PyMySQL
coverage
flask_sqlalchemy
pyyaml
nose2
flask-cors
flake8
tox
